+++ src/main/java/com/breakinblocks/plonk/client/ClientProxy.java
@@ -1,7 +1,9 @@
 package com.breakinblocks.plonk.client;
 
+import com.breakinblocks.plonk.client.command.CommandClientPlonk;
 import com.breakinblocks.plonk.client.registry.RegistryTESRs;
 import com.breakinblocks.plonk.common.CommonProxy;
+import net.minecraftforge.client.ClientCommandHandler;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 
 public class ClientProxy extends CommonProxy {
@@ -13,5 +15,6 @@ public class ClientProxy extends CommonProxy {
         super.init(event);
         RegistryTESRs.init();
         CLIENT_EVENTS.registerKeyBindings();
+        ClientCommandHandler.instance.registerCommand(new CommandClientPlonk());
     }
 }